robust to large M

+++ src/DFG.hpp
@@ -4,6 +4,6 @@
 #include "sc_utils.hpp"
 
 void OrderBySubCollection(SetCoverInput* sci, float p, vector<float>* p_pow_k, vector<vector<Set*>>* ktoCollection);
-vector<int>* DFG(SetCoverInput* sci, float p);
+vector<unsigned long>* DFG(SetCoverInput* sci, float p);
 
 #endif
